fix: readme, move functions in util.hpp to the library, add tips in yaml file

+++ test/include/load_yaml.hpp
@@ -52,7 +52,7 @@ bool BBS3DTest::load_config(const std::string& config) {
   min_scan_range = conf["min_scan_range"].as<double>();
   max_scan_range = conf["max_scan_range"].as<double>();
 
-  timeout_msec = conf["timeout_msec"].as<double>();
+  timeout_msec = conf["timeout_msec"].as<int>();
 
   use_gicp = conf["use_gicp"].as<bool>();
   return true;
